Add base EditorNode class
