+++ Assets/Scripts/Tools/Class Editor/ClassTreeNode.cs
@@ -4,9 +4,19 @@ using UnityEngine;
 
 namespace ClassEditor
 {
-    public class ClassTreeNode
+    public class ClassTreeNode : EditorNode
     {
         public bool unlocked;
         public bool taken;
+
+        public ClassTreeNode(Vector2 position) : base(position)
+        {
+
+        }
+
+        public override void Draw(Vector2 offset)
+        {
+
+        }
     }
 }
