+++ Assets/Scripts/Tools/Dialogue Graph/DialogueGraphNode.cs
@@ -5,23 +5,19 @@ using UnityEngine;
 namespace Dialogue
 {
     [System.Serializable]
-    public class DialogueGraphNode
+    public class DialogueGraphNode : EditorNode
     {
         private static readonly Vector2 nodeSize = new Vector2(200, 50);
 
         public string name;
         public string body;
-        public Vector2 position;
-        private Rect displayRect;
-        private bool isSelected = false;
 
-        public DialogueGraphNode(string name, Vector2 position)
+        public DialogueGraphNode(string name, Vector2 position) : base(position)
         {
             this.name = name;
-            this.position = position;
         }
 
-        public void Draw(Vector2 offset)
+        public override void Draw(Vector2 offset)
         {
             Vector2 pos = position - nodeSize / 2.0f;
             displayRect = new Rect(pos + offset, nodeSize);
@@ -29,39 +25,5 @@ namespace Dialogue
             string displayName = EditorUtils.TrimStringToFit(name, displayRect.width, GUI.skin.box);
             GUI.Box(displayRect, displayName, GUI.skin.button);
         }
-
-        public bool ProcessEvents(Event e)
-        {
-            switch (e.type)
-            {
-                case EventType.MouseDown:
-                    if (e.button == 0 && displayRect.Contains(e.mousePosition))
-                    {
-                        isSelected = true;
-                        GUI.changed = true;
-                    }
-                    break;
-                case EventType.MouseUp:
-                    if (e.button == 0)
-                    {
-                        isSelected = false;
-                        GUI.changed = true;
-                    }
-                    break;
-                case EventType.MouseDrag:
-                    if (isSelected)
-                    {
-                        position += e.delta;
-                        GUI.changed = true;
-                    }
-                    break;
-                default:
-                    break;
-            }
-
-            return isSelected;
-        }
-
-        public bool Contains(Vector2 position) => displayRect.Contains(position);
     }
 }
