+++ Assets/Scripts/Tools/EditorNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 145efbf5d0f39214684ef1f8f31bb030
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
